requirements.txt
openai==1.55.3
toml~=0.10.2
omegaconf~=2.3.0
google-cloud-texttospeech~=2.25.0
yt-dlp~=2025.2.19
elevenlabs~=1.52.0
pydub~=0.25.1
pysrt~=1.1.2
ffmpeg-python~=0.2.0
python-dotenv~=1.0.1